refactor(expense-documents): extract posting core into ExpenseDocumentLifecycleService (post/executePostBody/approve, verbatim)
Phase 2b of transactional-core decompose (HIGH RISK — JE posting). post,
executePostBody (shared inner-tx body), approve moved verbatim into the
lifecycle service; facade delegates. Facade sheds the 6 JE templates +
readNumberFlag + post/approve free-fn imports. +3 characterization tests
(CREDIT_NOTE post routing, PETTY_CASH allow-list throw quirk pinned,
CREDIT_NOTE auto-post). JE dispatch + state transitions byte-identical.
433 existing + 3 new green, tsc 0, lint 0.
